fixes #14434 - refactor tokens into concern from observer
rails-observers is still required via audited, but this removes the
usage in Foreman itself and unifies token-related code.
