+++ Gemfile
@@ -35,7 +35,6 @@ else
   gem 'net-ldap', '>= 0.8.0'
 end
 gem 'activerecord-session_store', '~> 0.1.1'
-gem 'rails-observers', '~> 0.1'
 gem 'protected_attributes', '~> 1.1.1'
 gem 'sprockets', '~> 3'
 gem 'sprockets-rails', '>= 2.3.3', '< 3'
